+++ pydidas/gui/managers/workflow_tree_edit_manager.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -19,23 +21,25 @@ with the editing Canvas.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["WorkflowTreeEditManager"]
 
+
 import time
 from functools import partial
 
-from qtpy import QtCore
 import numpy as np
+from qtpy import QtCore
 
-from ...core.constants import gui_constants
 from ...core import SingletonFactory
+from ...core.constants import gui_constants
 from ...plugins import PluginCollection
 from ...widgets.workflow_edit import PluginInWorkflowBox
-from ...workflow import WorkflowTree, PluginPositionNode
+from ...workflow import PluginPositionNode, WorkflowTree
+
 
 PLUGIN_COLLECTION = PluginCollection()
 TREE = WorkflowTree()
@@ -132,14 +136,10 @@ class _WorkflowTreeEditManager(QtCore.QObject):
 
         Raises
         ------
-        AttributeError
-            AttributeError is raised if the Qt drawing Canvas for the plugins
-            has not been defined yet.
+        Warning
+            If the Qt drawing Canvas for the plugins has not been defined yet.
         """
         _plugin = PLUGIN_COLLECTION.get_plugin_by_plugin_name(name)()
-        # Check if no node is active but the TREE already has nodes.
-        # If True, pydidas will select the last node as active to append the new
-        # node.
         _parent = (
             TREE.active_node if parent_node_id is None else TREE.nodes[parent_node_id]
         )
@@ -150,7 +150,7 @@ class _WorkflowTreeEditManager(QtCore.QObject):
         self.__create_widget(title if title is not None else name, _node_id)
         self.set_active_node(_node_id, force_update=True)
         if not self.qt_canvas:
-            raise Warning("No QtCanvas defined. Nodes added but cannot be displayed")
+            raise Warning("No QtCanvas defined. Nodes cannot be displayed")
         self.update_node_positions()
         self._check_consistency()
 
@@ -295,11 +295,13 @@ class _WorkflowTreeEditManager(QtCore.QObject):
             The node ID of the requested parent.
         """
         _plugin = TREE.nodes[calling_node].plugin
-        self.add_new_plugin_node(_plugin.plugin_name, parent_node_id=new_parent_node)
-        _dump = TREE.nodes[calling_node].dump()
-        for _key, _val in _dump["plugin_params"]:
-            TREE.active_node.plugin.set_param_value(_key, _val)
+        _param_vals = _plugin.get_param_values_as_dict()
+        self.add_new_plugin_node(
+            _plugin.plugin_name, parent_node_id=new_parent_node
+        )
+        TREE.active_node.plugin.set_param_values_from_dict(_param_vals)
         TREE.order_node_ids()
+        time.sleep(0.0005)
         self.update_from_tree()
         self._check_consistency()